+++ src/MatVecDouble.f90
@@ -6,7 +6,7 @@ module MatVecDouble
 contains
   type(DMat) function OuterProductD(a, b) result(c)
     type(DVec), intent(in) :: a, b
-    integer :: n, m
+    integer(4) :: n, m
     n = size(a%v)
     m = size(b%v)
     call c%ini(n,m)
@@ -16,7 +16,7 @@ contains
   type(DVec) function MVProductD(a, b) result(c)
     type(DMat), intent(in) :: a
     type(DVec), intent(in) :: b
-    integer :: m, k
+    integer(4) :: m, k
     m = size(a%m, 1)
     k = size(a%m, 2)
     if(size(a%m, 2) /= size(b%v)) then
@@ -30,7 +30,7 @@ contains
   type(DVec) function VMProductD(a, b) result(c)
     type(DMat), intent(in) :: b
     type(DVec), intent(in) :: a
-    integer :: m, k, n
+    integer(4) :: m, k, n
     m = size(b%m, 1)
     k = size(b%m, 2)
     n = size(a%v, 1)
