+++ src/Parameters.f90
@@ -1,4 +1,4 @@
 module LinAlgParameters
   implicit none
-  integer :: iseed(4) = (/3239, 4241, 1903, 1093/)
+  integer(4) :: iseed(4) = (/3239, 4241, 1903, 1093/)
 end module LinAlgParameters
